--- Assets/Scripts/Cards.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Cards
-{
-    private string type;
-
-    public Cards(string type)
-    {
-        this.type = type;
-    }
-
-}
