+++ Assets/Scripts/DataObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f57aa9f87baccdd4fb48edca64ef5f09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
