+++ Assets/Scripts/DataObject/Cards.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DataObject
+{
+    public class Cards
+    {
+        private string type;
+
+        public Cards(string type)
+        {
+            this.type = type;
+        }
+
+    }
+}
