+++ Assets/Scripts/DataObject/Creature.cs
@@ -1,0 +1,59 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DataObject
+{
+
+    public class Creature : MonoBehaviour
+    {
+
+        // entity var
+        private int health;
+        private int maxHealth;
+        private Vector3 position;
+        private bool invincible;
+
+        //entity fnc
+        public void dammage(int dammageValue)
+        {
+            if (!this.invincible)
+            {
+                this.health -= dammageValue;
+                if (this.health < 0)
+                {
+                    this.health = 0;
+                }
+
+            }
+
+        }
+
+        public void heal(int healValue)
+        {
+            this.health += healValue;
+            if (this.health > this.maxHealth)
+            {
+                this.health = this.maxHealth;
+            }
+        }
+
+        public void heal()
+        {
+            this.heal(this.maxHealth);
+        }
+
+        // Start is called before the first frame update
+        void Start()
+        {
+
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+
+        }
+    }
+}
+
