+++ Assets/Scripts/DataObject/Deck.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DataObject
+{
+    public class Deck
+    {
+        List<Cards> cards = new List<Cards>();
+
+
+    }
+}
