+++ Assets/Scripts/DataObject/PNJ.cs
@@ -1,0 +1,32 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+namespace DataObject
+{
+    public class PNJ : Creature
+    {
+
+        private IA ia;
+
+        private int goldReward;
+
+        Deck deck = new Deck();
+
+
+
+        // Start is called before the first frame update
+        void Start()
+        {
+
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+
+        }
+    }
+}
+
