+++ Assets/Scripts/DataObject/PNJ.cs.meta
