--- Assets/Scripts/Deck.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Deck {
-    List<Cards> cards = new List<Cards>();
-
-
-}
