--- Assets/Scripts/PNJ.cs
@@ -1,30 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-
-public class PNJ : Creature
-{
-
-    private IA ia;
-
-    private int goldReward;
-
-    Deck deck = new Deck();
-
-
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
-
